Move to geonode.core, rename view to forgot_username
