--- src/GeoNodePy/geonode/registration/urls.py
@@ -1,7 +1,0 @@
-from django.conf.urls.defaults import *
-
-
-urlpatterns = patterns('geonode.registration.views',
-    (r'^forgotname','forgotUsername'),
-    (r'', include('registration.urls')),
-)
