+++ src/GeoNodePy/geonode/templates/registration/login.html
@@ -13,9 +13,8 @@
 		  {{ form.as_ul }}
 		<li><input type="hidden" name="next" value="{{next}}" /></li>
 	    <li><input type="submit" value="Login"/></li>
-        <li><a href="{% url django.contrib.auth.views.password_reset %}">Forgot password?</a></li>
         <li><a href="{% url auth_password_reset %}">Forgot password?</a></li>
-        <li><a href="{% url geonode.registration.views.forgotUsername %}">Forgot username? </a></li>
+        <li><a href="{% url geonode.core.views.forgot_username %}">Forgot username? </a></li>
 	</ul>
   </form>
 </div>
