+++ src/GeoNodePy/geonode/urls.py
@@ -4,7 +4,6 @@ from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from geonode.sitemap import LayerSitemap, MapSitemap
 import geonode.proxy.urls
 import geonode.maps.urls
-import geonode.registration.urls
 
 # Uncomment the next two lines to enable the admin:
 from django.contrib import admin
@@ -37,6 +36,7 @@ urlpatterns = patterns('',
     # Accounts
     url(r'^accounts/ajax_login$', 'geonode.views.ajax_login', name='auth_ajax_login'),
     url(r'^accounts/ajax_lookup$', 'geonode.views.ajax_lookup', name='auth_ajax_lookup'),
+    url(r'^forgotname/$','geonode.core.views.forgot_username'),
     (r'^accounts/', include('registration.urls')),
     (r'^profiles/', include('profiles.urls')),
     (r'^avatar/', include('avatar.urls')),
